general update of the project + fix off-by-one with rolls
