+++ Cargo.toml
@@ -21,3 +21,4 @@ regex = "1"
 lazy_static = "1.4.0"
 evalexpr = "5.0.5"
 anyhow = "1.0"
+thiserror = "1.0"
